+++ CHANGELOG.php
@@ -1,11 +1,16 @@
+<!-- Changelog v5.3.0 -->
+1.) Updated to latest version of Foundation v5.3.0
+2.) Added Spanish Spanish Translation - Thanks to Alfonso Correas
+3.) Minimized font css a little more
+
 <!-- Changelog v5.2.3.1a -->
 1.) Seems in my haste to push out a fix for the Off Canvas issue, I forgot to change the version number in all of the files of WP-Forge. My apologies.
 
 <!-- Changelog v5.2.3.1 -->
 1.) Added "data-offcanvas" to line 13 of content-off_canvas.php - Seems this is needed by Off Canvas
 
-<!-- Changelog v5.2.3.1a -->
-1.) Updated to latest version of Foundation 5.2.3.1a
+<!-- Changelog v5.3.0 -->
+1.) Updated to latest version of Foundation 5.3.0
 2.) Added 'none' to the options for Background Position in the theme customizer.
 3.) Changed the default text of the menu fallback function
 4.) Updated normalize.css to 3.0.1 http://necolas.github.io/normalize.css/
@@ -42,7 +47,7 @@
 8.) Moved Modernizr to load after jQuery
 
 <!-- Changelog as of 04/06/2014 -->
-1.) Updated to Foundation v5.2.3.1a (released on Friday 04/04/2014) - View changes made by Foundation via commit https://github.com/zurb/foundation/commit/520ea11dfe818691dd70cdb5c5ed5907e5e64168
+1.) Updated to Foundation v5.3.0 (released on Friday 04/04/2014) - View changes made by Foundation via commit https://github.com/zurb/foundation/commit/520ea11dfe818691dd70cdb5c5ed5907e5e64168
 2.) Removed the default values in the Background Section of the theme customizer.
 3.) Added "background-size" section to Background Section of the theme customizer. This will allow users to set a full width image as the background of the site without a plugin. Read more about the "background-size" attribute http://www.w3schools.com/cssref/css3_pr_background-size.asp
 4.) Removed background color from style.css and foundation.css - The default color will now be white. This will allow the user to set the initial color from within the theme customizer.
@@ -79,7 +84,7 @@ theme customizer was set to "Excerpt".
 
 <!-- Changelog as of 03/18/2014 -->
 
-1.) Updated to latest version of Foundation, currently 5.2.3.1a
+1.) Updated to latest version of Foundation, currently 5.3.0
 2.) Changed the way Open-Sans is included into the theme. This will make it easier for anyone to remove the font or add their own if necessary.
 3.) Added support for Jetpack's infinite scroll
 4.) Added the actual search form to 404.php (I think it look better this way)