+++ footer.php
@@ -7,7 +7,7 @@
  *
  * @package WordPress
  * @subpackage WP_Forge
- * @since WP-Forge 5.2.3.1a
+ * @since WP-Forge 5.3.0
  */
 ?>
 	</section><!-- #main .wrapper -->