+++ js/foundation/foundation.tab.js
@@ -1,10 +1,13 @@
+/*
+@since WP-Forge 5.3.0
+*/
 ;(function ($, window, document, undefined) {
   'use strict';
 
   Foundation.libs.tab = {
     name : 'tab',
 
-    version : '5.2.3.1a',
+    version : '5.3.0',
 
     settings : {
       active_class: 'active',